wip: changing existing volume struct to docker volume helper; adding windows fix
