+++ pkg/volume/volume_test.go
@@ -4,54 +4,106 @@
 package volume
 
 import (
-	"context"
 	"testing"
-
-	"github.com/Azure/acr-builder/pkg/procmanager"
-
-	"github.com/Azure/acr-builder/util"
 )
 
-func TestNewVolume(t *testing.T) {
+func TestVolumeValidate(t *testing.T) {
 	tests := []struct {
-		name               string
-		expectedCreateArgs []string
-		expectedDeleteArgs []string
+		volumemount *Volume
+		shouldError bool
 	}{
 		{
-			"foo",
-			[]string{"docker", "volume", "create", "--name", "foo"},
-			[]string{"docker", "volume", "rm", "foo"},
+			nil,
+			false,
+		},
+		{
+			&Volume{
+				Name: "",
+				Values: []map[string]string{
+					{
+						"a": "this is a test",
+					},
+				},
+			},
+			true,
+		},
+		{
+			&Volume{
+				Name:   "test",
+				Values: []map[string]string{},
+			},
+			true,
+		},
+		{
+			&Volume{
+				Name:   "",
+				Values: []map[string]string{},
+			},
+			true,
+		},
+		{
+			&Volume{
+				Name: "test",
+				Values: []map[string]string{
+					{
+						"a": "this is a test",
+					},
+				},
+			},
+			false,
+		},
+		{
+			&Volume{
+				Name: "test123-_",
+				Values: []map[string]string{
+					{
+						"a": "this is a test",
+					},
+				},
+			},
+			false,
+		},
+		{
+			&Volume{
+				Name: "test/.",
+				Values: []map[string]string{
+					{
+						"a": "this is a test",
+					},
+				},
+			},
+			true,
+		},
+		{
+			&Volume{
+				Name: "test",
+				Values: []map[string]string{
+					{
+						"": "this is a test",
+					},
+				},
+			},
+			true,
+		},
+		{
+			&Volume{
+				Name: "test",
+				Values: []map[string]string{
+					{
+						"a": "",
+					},
+				},
+			},
+			false,
 		},
 	}
-	procManager := procmanager.NewProcManager(true)
-
 	for _, test := range tests {
-		vol := NewVolume(test.name, procManager)
-		if vol.Name != test.name {
-			t.Fatalf("Expected volume name: %s but got %s", test.name, vol.Name)
+		err := test.volumemount.Validate()
+		if test.shouldError && err == nil {
+			t.Fatalf("Expected volume mount: %v to error but it didn't", test.volumemount)
 		}
-		if actual := vol.getDockerCreateArgs(); !util.StringSequenceEquals(actual, test.expectedCreateArgs) {
-			t.Fatalf("Expected %v as the create args, but got %v", test.expectedCreateArgs, actual)
-		}
-		if actual := vol.getDockerRmArgs(); !util.StringSequenceEquals(actual, test.expectedDeleteArgs) {
-			t.Fatalf("Expected %v as the delete args, but got %v", test.expectedCreateArgs, actual)
-		}
-
-		out, err := vol.Create(context.Background())
-		if err != nil {
-			t.Fatalf("Unexpected err during volume creation: %v", err)
-		}
-		if out != "" {
-			t.Fatalf("Unexpected output from volume creation: %s", out)
-		}
-
-		out, err = vol.Delete(context.Background())
-		if err != nil {
-			t.Fatalf("Unexpected err during volume deletion: %v", err)
-		}
-		if out != "" {
-			t.Fatalf("Unexpected output from volume deletion: %s", out)
+		if !test.shouldError && err != nil {
+			t.Fatalf("volume mount: %v shouldn't have errored, but it did; err: %v", test.volumemount, err)
 		}
 	}
 }
